Encode long stats numbers to json properly
git-svn-id: https://svn.smr.arpa.emr.it/arkimet/arkimet/trunk@2258 9b5092fc-ae44-42f6-b796-0d08dad2e17b

+++ arki/summary/stats.cc
@@ -143,8 +143,8 @@ void Stats::serialiseLocal(Emitter& e, const Formatter* f) const
             e.add("e"); reftimeMerger.begin->serialiseList(e);
         }
     }
-    e.add("c", (int)count);
-    e.add("s", (int)size);
+    e.add("c", (long long int)count);
+    e.add("s", (long long int)size);
 }
 
 UItem<Stats> Stats::decodeMapping(const emitter::memory::Mapping& val)
